pyproject.toml
[project]
name = "collab-tasks-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "JeissonRuiz02",email = "jeissonruizdev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "google-cloud-firestore (>=2.20.1,<3.0.0)",
    "google-cloud-pubsub (>=2.29.0,<3.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "firestore (>=0.0.8,<0.0.9)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4)",
    "fastapi-users (>=14.0.1,<15.0.0)",
    "auth (>=0.5.3,<0.6.0)",
    "bcrypt (>=4.0.1)"
]

[tool.poetry]
packages = [{include = "collab_tasks_backend", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
